fix: forward environment overrides to popup sessions (#6)
As mentioned in <https://github.com/loichyan/tmux-toggle-popup/pull/5#issuecomment-2415237970>, environment overrides are not set properly in opened popups. This is because commands actually start inside a tmux session, which is typically running in a standalone tmux instance, so environment overrides in `display-popup -e KEY=value` are not set in the attached tmux session unless we explicitly forward them to `new-session -e KEY=value`.