feat!: support switch popups when toggled (#21)
If the user defines many different popup key bindings, it is often the case that forcing opened popups to close or popup-in-popups is not the desired behavior to deal with nested toggle calls. This PR introduces a new toggle mode, `switch`, which allows for quick switching between open popups. Specifically, in the `switch` toggle mode, when a different popup is toggled, `@popup-toggle` will reuse the currently opened popup window and switch to the newly specified one.

BREAKING CHANGE: 

1. `@popup-toggle --force` is replaced with `--toggle-mode=force-close`. 
2. tmux variable `#{@popup_name}` in `@popup-id-format` is replaced with `{popup_name}`.